requirements.txt
matplotlib>=3.3.4
Pillow>=9.0.0
setuptools>=52.0.0.post20210125
